+++ Objects/object.c
@@ -5,7 +5,6 @@
 #include "pycore_pystate.h"
 #include "pycore_context.h"
 #include "frameobject.h"
-#include "interpreteridobject.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -1807,7 +1806,6 @@ _PyTypes_Init(void)
     INIT_TYPE(&PySeqIter_Type, "sequence iterator");
     INIT_TYPE(&PyCoro_Type, "coroutine");
     INIT_TYPE(&_PyCoroWrapper_Type, "coroutine wrapper");
-    INIT_TYPE(&_PyInterpreterID_Type, "interpreter ID");
     return _Py_INIT_OK();
 
 #undef INIT_TYPE
